AlignWithClosestGridTile extension for EntityCoordinates - Fixes component exception when adding a timer component to a deleted entity.
